pipelines/account_opening/requirements.txt
pandas==1.5.3
pyodbc==4.0.34
python-dotenv==1.0.1
